Move getNodeSourceText from ast/locations.ts to text/source-text.ts
**Rationale:** getNodeSourceText is a pure text-extraction utility that maps an
AST node's start/end boundaries to a slice of the source text. It belongs in
src/core/src/text alongside other text-processing helpers (validateSourceText,
splitLines, dominantLineEnding, string/comment scanners) rather than in
src/core/src/ast/locations.ts, which owns structural location metadata
(getNodeStartIndex, cloneLocation, etc.).

**Changes:**
- src/core/src/ast/locations.ts: Removed getNodeSourceText function and export
- src/core/src/text/source-text.ts: Added getNodeSourceText with its JSDoc, importing
  the required getNodeStartIndex/getNodeEndIndex from ../ast/locations.js
- src/core/src/ast/node-helpers/index.ts: Updated re-export path to point to
  ../../text/source-text.js instead of ../locations.js

All imports resolved, build passes, and lint is clean.
