+++ src/core/src/ast/node-helpers/index.ts
@@ -1,4 +1,4 @@
-export { getNodeSourceText } from "../locations.js";
+export { getNodeSourceText } from "../../text/source-text.js";
 export * from "./basics.js";
 export * from "./identifiers.js";
 export * from "./literals-and-operators.js";
